Update buttons field to support more buttons
Changed buttons from uint8_t to uint16_t to support up to 16 buttons.

+++ Core/Inc/gamepad_hid.h
@@ -5,7 +5,7 @@
 
 // Struct 7 bytes: Trình tự phải giống y hệt thứ tự mô tả trong Report Descriptor
 typedef struct __attribute__((packed)) {
-    uint8_t  buttons;    // Bit 0->7: 8 nút bấm
+    uint16_t buttons;    // Dùng 16 bits để chứa tối đa 16 nút
     uint8_t  left_x;     // 0=Left, 128=Center, 255=Right
     uint8_t  left_y;     // 0=Up, 128=Center, 255=Down
     uint8_t  right_x;    // 0=Left, 128=Center, 255=Right
